@@ -1,6 +1,6 @@
 {
   "name": "react-froala-wysiwyg",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "description": "React component for Froala WYSIWYG HTML rich text editor.",
   "main": "index.js",
   "types": "./lib/index.d.ts",
@@ -22,52 +22,44 @@
     "react-dom": "~0.14 || ^15.0.0 || ^16.0.0 || ^17.0.0 || ^18.0.0 || ^19.0.0"
   },
   "dependencies": {
-    "create-react-class": "^15.5.2",
-    "froala-editor": "4.3.1"
+    "froala-editor": "4.4.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-decorators": "^7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
-    "@babel/plugin-proposal-function-sent": "^7.0.0",
-    "@babel/plugin-proposal-json-strings": "^7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-syntax-import-meta": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.12.13",
-    "autoprefixer": "^8.3.0",
+    "@babel/cli": "^7.25.9",
+    "@babel/core": "^7.26.0",
+    "@babel/plugin-proposal-decorators": "^7.25.9",
+    "@babel/plugin-proposal-function-sent": "^7.25.9",
+    "@babel/plugin-proposal-throw-expressions": "^7.25.9",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.25.9",
+    "autoprefixer": "^10.4.20",
     "babel-loader": "^8.0.0",
-    "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^0.28.11",
+    "copy-webpack-plugin": "^12.0.2",
+    "core-js": "^3.39.0",
+    "css-loader": "^7.1.2",
     "exports-loader": "^0.6.2",
-    "file-loader": "^1.1.11",
-    "highlight.js": "^9.3.0",
-    "imports-loader": "^0.6.4",
-    "jsdom": "^9.0.0",
-    "mocha": "^3.0.0",
+    "file-loader": "^6.2.0",
+    "highlight.js": "^11.10.0",
+    "imports-loader": "^5.0.0",
+    "jsdom": "^25.0.1",
+    "mocha": "^10.8.2",
     "mock-require": "^1.3.0",
-    "nightwatch": "^0.9.6",
-    "postcss-loader": "^2.1.4",
+    "nightwatch": "^3.9.0",
+    "postcss-loader": "^8.1.1",
     "raw-loader": "^0.5.1",
-    "react": "^16.0.0",
-    "react-addons-test-utils": "^15.0.0",
-    "react-dom": "^16.0.0",
-    "react-highlight": "^0.9.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-highlight": "^0.15.0",
     "sinon": "^1.17.4",
     "style-loader": "^0.21.0",
-    "url-loader": "^0.5.5",
-    "webpack": "^4.32.1",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.3.0"
+    "url-loader": "^4.1.1",
+    "webpack": "^5.96.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0"
   },
   "scripts": {
-    "build": "webpack-cli && webpack-cli -p",
-    "demo": "cd demo && webpack-cli && \"../node_modules/.bin/webpack-dev-server\" --disableHostCheck=true --port 4000 --host 0.0.0.0 --content-base dist/",
+    "build": "webpack-cli --mode=development && webpack-cli --mode=production",
+    "demo": "cd demo && webpack-cli --mode=development && \"../node_modules/.bin/webpack-dev-server\" --mode development --allowed-hosts all --port 4000 --host 0.0.0.0 --static dist/",
     "prepublishOnly": "npm run build && bash lib/copy_bundles.sh"
   },
   "repository": {
