@@ -7,7 +7,7 @@ import 'file-loader?name=[name].[ext]!./init_on_input.html';
 
 import FroalaEditorInput from 'react-froala-wysiwyg/FroalaEditorInput';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 
 // Render Froala Editor component.
 class EditorComponent extends React.Component {
@@ -44,5 +44,6 @@ class EditorComponent extends React.Component {
   }
 }
 
-ReactDOM.render(<EditorComponent/>, document.getElementById('editor'));
+const root = ReactDOM.createRoot(document.getElementById('editor'));
+root.render(<EditorComponent/>);
 
