@@ -12,7 +12,7 @@ module.exports = {
     init_on_button: './src/init_on_button.jsx',
     init_on_link: './src/init_on_link.jsx',
     init_on_input: './src/init_on_input.jsx',
-	custombutton: './src/custombutton.jsx'
+    custombutton: './src/custombutton.jsx'
   },
 
   optimization: {
@@ -29,9 +29,10 @@ module.exports = {
             cacheDirectory: true,
               presets: [
                 ['@babel/preset-env', {
-                  'targets': { 
+                  'targets': {
                     "ie": "11"
                   },
+                  "corejs": 3,
                   "useBuiltIns": "entry"
                 }],
                 '@babel/preset-react']
@@ -41,7 +42,12 @@ module.exports = {
         test: /\.css$/,
         use: [
           'style-loader',
-          'css-loader'
+          {
+            loader: 'css-loader',
+            options: {
+              esModule: false
+            }
+          }
         ]
       },
       {
@@ -81,6 +87,11 @@ module.exports = {
 
     }),
 
-    new CopyWebpackPlugin([{ from: './src/index.html'}, {from: './src/image.jpg'} ])
+    new CopyWebpackPlugin({
+      patterns: [
+        { from: './src/index.html' },
+        { from: './src/image.jpg' }
+      ]
+    })
   ]
 };
